This new implementation really simplifies the start/end tag classes
